Completed ui design
