+++ App.js
@@ -30,6 +30,7 @@ export default function App(props) {
         await Font.loadAsync({
           ...Ionicons.font,
           'space-mono': require('./assets/fonts/SpaceMono-Regular.ttf'),
+          'Roboto_medium': require("native-base/Fonts/Roboto_medium.ttf")
         });
       } catch (e) {
         // We might want to provide this error information to an error reporting service
