+++ package.json
@@ -24,6 +24,7 @@
     "expo-font": "~8.0.0",
     "expo-web-browser": "~8.0.0",
     "express": "^4.17.1",
+    "native-base": "^2.13.8",
     "nodemon": "^2.0.2",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
